+++ models/sale.js
@@ -24,15 +24,15 @@ module.exports = function(sequelize, DataTypes) {
       allowNull: false
     },
     currency_quantity: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(30,20),
       allowNull: false
     },
     currency_charge: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(30,20),
       allowNull: false
     },
     currency_total: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(30,20),
       allowNull: false
     }
   }, {
